+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Releases
 
+### v0.1.7 (2025-11-17)
+
 - **Added:** `DOC013` to check that the keys of the sections of the documentation chain are the common ones.
 - **Added:** `DOC014` to check if docstring summary fits on a single line.